change the data type of rectangle from int to float for more precision, considered making it a typename, decided against it

+++ debug.cpp
@@ -5,7 +5,7 @@
 
 int main(int argc, char *argv[])
 {
-    int points[5][2] = {{20,6},{99,70},{20,60},{4,70},{70,63}}, bbs[5][2][2], data[5] = {1,2,3,4,5};
+    float points[5][2] = {{20,6},{99,70},{20,60},{4,70},{70,63}}, bbs[5][2][2], data[5] = {1,2,3,4,5};
     for(int i = 0; i < 5; i++){
         int idx1 = i, idx2 = (i + 1) % 5;
         if(points[idx1][0] < points[idx2][0]){
@@ -30,13 +30,13 @@ int main(int argc, char *argv[])
     // }
     // printf("--------------------\n");
 
-    struct STRTREE<int,2> tree((int***)bbs, data, 5);
+    struct STRTREE<float,2> tree((float***)bbs, data, 5);
     tree.print_tree();
 
     int q[2][2] = {{20,30},{60,40}};
-    std::vector<struct Branch<int> *> *res = tree.retrieve_all_leaves();
+    std::vector<struct Branch<float> *> *res = tree.retrieve_all_leaves();
     for(auto b : *res){
-        printf("%d\n",b->data);
+        printf("%f\n",b->data);
     }
 
     delete res;
